Return Result<false> on thick device
The function is_thin_device was raising Result<Err> when the device
wasn't a thinly provisioned device.  This prevents blk-archive from
working on devices that aren't thinly provisioned.  Documentation
indicates that this is a use case, so change to return Result<false>
when device isn't thin so that we proceed to use thick_packer.

Signed-off-by: Tony Asleson <tasleson@redhat.com>