Add digest for stored streams
Today the only way you have to check if a stream is correct is by
comparing the stream with the device or file that the stream
orginated from.  This change adds a blake2 512bit digest to the
stream config.toml.  This can be used to validate what we unpack
matches what we packed.  This change also adds a "validate" command
with a sub command of stream.

Note: This change only applies to files or thick devices.  Delta or
thin functionality needs to yet be included.

The digest in the config.yaml matches the output of b2sum, to make
it easier to validate that blk-archive is calculating the digest
correctly, eg.

 # b2sum /dev/sdb

Matches what we have in the config.yaml for an archive of the same
device.

Signed-off-by: Tony Asleson <tasleson@redhat.com>