Introduce an abstraction for data slab handling
This change introduces an abstraction for pack and unpack
operations. After creating an instance of Db, use data_add
for packing and data_get for unpacking.

The Db struct encapsulates a Cuckoo filter and an LruCache
for lookups during data_add, as well as a separate LruCache
for data_get operations.

Signed-off-by: Tony Asleson <tasleson@redhat.com>