Keep LRU link_list actions in the lru_ functions
The evict and insert branch of the push() function sets the LRUs tail
to entries[tail].next. Then it calls lru_del_() which would do the same
thing, except that the push() code already did it. There's no need to
do it in both places, and it's better to keep the list modification
code contained, so remove the unnecessary code from push().