Avoid unnecesary work for updating LRU tail and head entries
Joining lru_del() and lru_add() into one function, lru_update(), allows
us to easily avoid doing unnecessary work in the case where you are
pushing a value already at the head of the LRU (no work is necessary) or
evicting a value from the tail (only updating head and tail are necessary).

Since lru_del_() removes an entry from the linked list but leaves it
in the Vec and lru_add() takes an unlinked entry in the Vec and adds it
to the head of the linked list, there is never a reason to call only one
of these functions.