Finish off slab when stream is saved
After a client sends us a stream file to save we will complete
and data waiting in a slab.

Note: This is an incomplete fix as there is still a race condition
between writing a chunk and being able to retrieve it.  This is
due to the delay it takes to write a chunk as it's done in a
thread pool.

Signed-off-by: Tony Asleson <tasleson@redhat.com>