[thin_metadata] drop get_thin_name and get_thin_details_
get_thin_details() is exactly the same as calling get_thin_name() and
then get_thin_details_(). Use it instead, in read_thin_delta(). If
new_thin is not a dm_device, this change will cause the table of
old_thin to get parsed unnecessarily, but this isn't much extra work. In
exchange, there is less code duplication, and if both devices are dm
device but one isn't a thin device, the error message will now report
which argument was wrong.