[unpack] Remove unused code
Today when you compile 'main' branch you get the following warnings

warning: method `handle_fill` is never used
  --> src/unpack.rs:30:8
   |
29 | trait UnpackDest {
   |       ---------- method in this trait
30 |     fn handle_fill(&mut self, byte: u8, len: u64) -> Result<()>;
   |        ^^^^^^^^^^^
   |
   = note: `#[warn(dead_code)]` on by default

warning: methods `fill`, `handle_fill_unprovisioned`, and `handle_fill_provisioned` are never used
   --> src/unpack.rs:255:8
    |
211 | impl ThinDest {
    | ------------- methods in this implementation
...
255 |     fn fill(&mut self, byte: u8, len: u64) -> Result<()> {
    |        ^^^^
...
270 |     fn handle_fill_unprovisioned(&mut self, byte: u8, len: u64) -> Result<()> {
    |        ^^^^^^^^^^^^^^^^^^^^^^^^^
...
280 |     fn handle_fill_provisioned(&mut self, byte: u8, len: u64) -> Result<()> {
    |        ^^^^^^^^^^^^^^^^^^^^^^^

warning: `blk-archive` (lib) generated 2 warnings
    Finished `release` profile [optimized + debuginfo] target(s) in 3.35s

This commit removes the code that is not being used and preserves what
is being executed today.  However, these warning could also be removed
and the excluded code to be included in execution if we simply called
the 'handle_fill' method in the 'unpack_entry' when encountering a
Fill entry.  As that would be executing code that hasn't been exercised
in a considerable amount of time I've elected to remove it.

Signed-off-by: Tony Asleson <tasleson@redhat.com>