[stats] Don't use file sizes for data_written
Instead of utilizing the sizes of the data file, we'll
keep a running talley of each while processing the file.  This approach
will continue to work when we introduce a client/server arch. with
multiple clients writing to the same archive concurrently.

Note: We're removing the hash data written and not including meta data
sizes in returned values.

Signed-off-by: Tony Asleson <tasleson@redhat.com>